Funcionality to create "curves" for models LZCN, RZCN, R3CN
Expose in hwmon sysfs fancurve_indices, returns the current indices of
the fan table that is used by the bios. Also allows to be modified when
in Custom powermode (255)

Read
cat /path/to/hwmon/fancurve_indices
1,2,3,4,5,6,7,8,9,10

Write
echo "2,2,3,4,5,6,7,8,9,10" > /path/to/hwmon/fancurve_indices

It will use the rpm's from indice 2 in indice 1, similar to
what the Vantage App does. Values allowed from 1 to 10.

It takes in consideration the /path/to/hwmon/auto_points_defaults.

example: If you are in custom mode "balanced-performance" (255), and you
set to use the default fancurve of quiet powermode :

echo 1 > /path/to/hwmon/auto_points_defaults
cat /path/to/hwmon/fancurve_indices
1,2,3,4,5,6,7,8,9,10
echo "1,2,3,4,4,4,7,8,9" > /path/to/hwmon/fancurve_indices

will create a curve with the rpm's of quiet powermode.